requirements.txt
decorator==4.0.4
Django==1.8.5
django-braces==1.8.1
django-extensions==1.5.7
django-file-resubmit==0.4.2
django-paypal==0.2.7
invoke==0.11.1
ipython==4.0.0
ipython-genutils==0.1.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
Pillow==3.0.0
ptyprocess==0.5
python-dateutil==2.4.2
requests==2.8.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.0.0
Werkzeug==0.10.4
wheel==0.24.0
django-watson==1.1.8
isbnlib==3.5.7